Move test helpers into separate directory - rename types for export from package
